refactor: Make RPC detection methods function receivers
