+++ types/rest.go
@@ -1,4 +1,4 @@
-package rpcdetect
+package types
 
 import (
 	"fmt"
@@ -6,43 +6,25 @@ import (
 	"slices"
 
 	"google.golang.org/protobuf/proto"
-
-	"github.com/pokt-network/shannon-sdk/types"
 )
 
 var (
 	// defaultRESTErrorReply is the default REST error reply to be sent if the
 	// formatRESTError function fails to format the appropriate one.
-	defaultRESTErrorReply   *types.POKTHTTPResponse
+	defaultRESTErrorReply   *POKTHTTPResponse
 	defaultRESTErrorReplyBz []byte
 )
 
-// init initializes the default REST error reply. This function is called before
-// the main function and panics if it fails to marshal the default REST error reply,
-// making the program exit early.
+// init initializes the package level variables for the REST error reply.
 func init() {
-	// Initialize the default REST error reply
-	header := &types.Header{
-		Key:    contentTypeHeaderKey,
-		Values: []string{"text/plain"},
-	}
-	headers := map[string]*types.Header{contentTypeHeaderKey: header}
-
-	defaultRESTErrorReply = &types.POKTHTTPResponse{
-		StatusCode: http.StatusInternalServerError,
-		Header:     headers,
-		BodyBz:     []byte(`Internal error`),
-	}
-
-	var err error
-	defaultRESTErrorReplyBz, err = proto.Marshal(defaultRESTErrorReply)
-	if err != nil {
-		panic(err)
-	}
+	// Initialize the default REST error reply and panic if it fails. This is done
+	// to make the program exit early if the default REST error reply fails to be
+	// marshaled.
+	initDefaultRESTErrorReply()
 }
 
 // isREST checks if the given POKTHTTPRequest is a REST request.
-func isREST(poktRequest *types.POKTHTTPRequest) bool {
+func (poktRequest *POKTHTTPRequest) isREST() bool {
 	// Since a REST request could have an empty body, can be of any method, and
 	// can have any content type, we can't use those to determine if a request is
 	// a REST request.
@@ -54,35 +36,34 @@ func isREST(poktRequest *types.POKTHTTPRequest) bool {
 
 // formatRESTError formats the given error into a POKTHTTPResponse and its
 // corresponding byte representation.
-func formatRESTError(
+func (poktRequest *POKTHTTPRequest) formatRESTError(
 	err error,
-	poktRequest *types.POKTHTTPRequest,
 	isInternal bool,
-) (*types.POKTHTTPResponse, []byte) {
+) (*POKTHTTPResponse, []byte) {
 	errorMsg := err.Error()
 	statusCode := http.StatusBadRequest
 	if isInternal {
-		errorMsg = "Internal error"
+		errorMsg = defaultErrorMessage
 		statusCode = http.StatusInternalServerError
 	}
 
-	contentTypeHeaderValues := []string{"text/plain"}
+	contentTypeHeaderValues := []string{contentTypeHeaderValueText}
 	if contentTypeHeader, ok := poktRequest.Header[contentTypeHeaderKey]; ok {
 		contentTypeHeaderValues = contentTypeHeader.Values
 	}
 
 	responseBodyBz := []byte(errorMsg)
-	if slices.Contains(contentTypeHeaderValues, "application/json") {
+	if slices.Contains(contentTypeHeaderValues, contentTypeHeaderValueJSON) {
 		responseBodyBz = []byte(fmt.Sprintf(`{"error": "%s"}`, errorMsg))
 	}
 
-	header := &types.Header{
+	header := &Header{
 		Key:    contentTypeHeaderKey,
 		Values: contentTypeHeaderValues,
 	}
-	headers := map[string]*types.Header{contentTypeHeaderKey: header}
+	headers := map[string]*Header{contentTypeHeaderKey: header}
 
-	poktResponse := &types.POKTHTTPResponse{
+	poktResponse := &POKTHTTPResponse{
 		StatusCode: uint32(statusCode),
 		Header:     headers,
 		BodyBz:     responseBodyBz,
@@ -94,4 +75,28 @@ func formatRESTError(
 	}
 
 	return poktResponse, poktResponseBz
+}
+
+// initDefaultRESTErrorReply initializes the default REST error reply.
+// This function is called before the main function and panics if it fails to
+// marshal the default REST error reply, making the program exit early.
+func initDefaultRESTErrorReply() {
+	// Initialize the default REST error reply
+	header := &Header{
+		Key:    contentTypeHeaderKey,
+		Values: []string{contentTypeHeaderValueText},
+	}
+	headers := map[string]*Header{contentTypeHeaderKey: header}
+
+	defaultRESTErrorReply = &POKTHTTPResponse{
+		StatusCode: http.StatusInternalServerError,
+		Header:     headers,
+		BodyBz:     []byte(defaultErrorMessage),
+	}
+
+	var err error
+	defaultRESTErrorReplyBz, err = proto.Marshal(defaultRESTErrorReply)
+	if err != nil {
+		panic(err)
+	}
 }
